pyproject.toml
[tool.poetry]
name = "dbnotebook"
version = "0.2.0"
description = "Local RAG chatbot with notebook management and vision capabilities"
authors = ["Võ Đình Đạt <vodinhdat03@gmail.com>"]
readme = "README.md"
packages = [{include = "dbnotebook"}]

[tool.poetry.scripts]
dbnotebook = "dbnotebook.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
einops = "^0.7.0"
python-dotenv = "^1.0.1"
sentence-transformers = "^2.5.1"
pydantic = "^2.6.4"
llama-index-embeddings-huggingface = "^0.1.4"
llama-index-llms-openai = "^0.1.12"
llama-index-llms-anthropic = "^0.1.15"
llama-index-llms-gemini = "^0.1.11"
llama-index-vector-stores-postgres = "^0.1.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pgvector = "^0.2.4"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
llama-index-llms-ollama = "^0.1.2"
llama-index-readers-file = "^0.1.11"
llama-index = "^0.10.22"
transformers = "^4.39.1"
torch = "*"
llama-index-callbacks-wandb = "^0.1.2"
llama-index-retrievers-bm25 = "^0.1.3"
pytest = "^8.2.0"
pymupdf = "^1.24.3"
google-generativeai = "^0.3.0"
flask = "^3.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
